pyproject.toml
[tool.poetry]
name = "prosnet-prefect-pipelines"
version = "0.1.0"
description = ""
authors = ["Matthias Schlögl <matthias.schloegl@oeaw.ac.at>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "prosnet-prefect-pipelines" }
]

[tool.poetry.dependencies]
python = "^3.11"
typesense = "^0.19.0"
sparqlwrapper = "^2.0.0"
prefect = {git = "https://github.com/acdh-oeaw/prefect.git", branch = "v2.20.3"}
gitpython = "^3.1.43"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
